# Conflicts:
#	.github/workflows/release.yml
#	gradle.properties
#	src/main/java/io/github/omegabird113/cmd_delete/command/CommandCreationUtils.java
#	src/main/java/io/github/omegabird113/cmd_delete/config/fileio/MappingsJSONManager.java
#	src/main/java/io/github/omegabird113/cmd_delete/mappings/NavMappings.java
#	src/main/java/io/github/omegabird113/cmd_delete/mixin/EditBoxMixin.java
#	src/main/java/io/github/omegabird113/cmd_delete/mixin/MultilineTextFieldMixin.java
#	src/main/java/io/github/omegabird113/cmd_delete/mixin/SignEditScreenMixin.java
#	src/main/resources/fabric.mod.json

+++ src/main/java/io/github/omegabird113/cmd_delete/command/MappingsInfoCollectionUtils.java
@@ -3,7 +3,7 @@ package io.github.omegabird113.cmd_delete.command;
 import io.github.omegabird113.cmd_delete.config.data.MappingsIdResolutionUtils;
 import io.github.omegabird113.cmd_delete.config.fileio.MappingsJSONManager;
 import io.github.omegabird113.cmd_delete.mappings.MappingsState;
-import io.github.omegabird113.cmd_delete.mappings.Os;
+import io.github.omegabird113.cmd_delete.utils.Os;
 import org.jetbrains.annotations.Contract;
 import org.jspecify.annotations.NonNull;
 
@@ -18,7 +18,7 @@ public final class MappingsInfoCollectionUtils {
 
     @Contract(pure = true)
     public static @NonNull String getInfoFrom(@NonNull MappingsState mappingsState, boolean includeDescription) {
-        final float coverage = mappingsState.mappings().getCoverage();
+        final double coverage = mappingsState.mappings().getCoverage();
 
         String displayName = "";
         String description = "";
